+++ .gitignore
@@ -14,6 +14,7 @@ Icon
 
 # virtualenv
 .virtualenv/
+venv/
 
 
 # Intellij
