+++ .travis.yml
@@ -3,6 +3,7 @@ python:
 - '3.6'
 - '3.7'
 - '3.8'
+- '3.9'
 before_install: sudo apt-get install unzip
 before_script:
   - export TFVER=0.13.4
