+++ python_terraform/tfstate.py
@@ -1,19 +1,20 @@
 import json
 import logging
 import os
+from typing import Dict, Optional
 
 logger = logging.getLogger(__name__)
 
 
 class Tfstate:
-    def __init__(self, data=None):
-        self.tfstate_file = None
+    def __init__(self, data: Optional[Dict[str, str]] = None):
+        self.tfstate_file: Optional[str] = None
         self.native_data = data
         if data:
             self.__dict__ = data
 
     @staticmethod
-    def load_file(file_path):
+    def load_file(file_path: str) -> "Tfstate":
         """Read the tfstate file and load its contents.
 
         Parses then as JSON and put the result into the object.
